Fix bug in validity check if vessel graph skeleton voxels
Also fix warnings in several places and slightly reorganize VolumeMask creation

+++ voreen/custommodules/vesseltopology/datastructures/vesselgraph.h
@@ -111,7 +111,7 @@ class VGNodeID {
 public:
     const static VGNodeID INVALID;
     VGNodeID()
-        : internal_(-1)
+        : internal_((uint32_t)-1)
     {
     }
 
@@ -141,7 +141,7 @@ class VGEdgeID {
 public:
     const static VGEdgeID INVALID;
     VGEdgeID()
-        : internal_(-1)
+        : internal_((uint32_t)-1)
     {
     }
     VGEdgeID(uint32_t i)
